requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   google-genai
    #   httpx
blinker==1.9.0
    # via flask
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.6
    # via requests
click==8.3.1
    # via flask
cryptography==46.0.5
    # via google-auth
distro==1.9.0
    # via google-genai
flask==3.1.3
    # via
    #   polish-case-visualizer (pyproject.toml)
    #   flask-cors
flask-cors==6.0.2
    # via polish-case-visualizer (pyproject.toml)
google-auth==2.49.1
    # via google-genai
google-genai==1.68.0
    # via polish-case-visualizer (pyproject.toml)
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via google-genai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
pyasn1==0.6.3
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via google-genai
pydantic-core==2.41.5
    # via pydantic
requests==2.32.5
    # via
    #   google-auth
    #   google-genai
sniffio==1.3.1
    # via google-genai
tenacity==9.1.4
    # via google-genai
typing-extensions==4.15.0
    # via
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via requests
websockets==16.0
    # via google-genai
werkzeug==3.1.7
    # via
    #   flask
    #   flask-cors
